LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-line-clamp-003.tentative-expected.txt




































PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 25
PASS .target > * 4
PASS .target > * 5
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="65"><span></span></div>
offsetLeft expected 65 but got 25
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 25
PASS .target > * 12
PASS .target > * 13
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 25
PASS .target > * 16
PASS .target > * 17
PASS .target > * 18
FAIL .target > * 19 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 25
PASS .target > * 20
PASS .target > * 21
PASS .target > * 22
FAIL .target > * 23 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 25
PASS .target > * 24
